+++ releaf-core/lib/generators/dummy/templates/controllers/admin/books_controller.rb
@@ -1,6 +1,2 @@
 class Admin::BooksController < Releaf::BaseController
-  def setup
-    super
-    @searchable_fields = [:title]
-  end
 end