+++ releaf-i18n_database/app/builders/releaf/i18n_database/translations/index_builder.rb
@@ -36,5 +36,9 @@ module Releaf::I18nDatabase::Translations
       url = url_for(action: :edit, search: params[:search])
       button(t("Edit"), "edit", class: "primary", href: url)
     end
+
+    def text_search_available?
+      true
+    end
   end
 end