+++ releaf-permissions/spec/features/roles_spec.rb
@@ -5,6 +5,16 @@ feature "Roles management", js: true do
     @role = Releaf::Permissions::Role.first
   end
 
+  scenario "Role search", focus: true do
+    create(:admin_role, name: 'super role')
+    visit releaf_permissions_roles_path
+    expect(page).to have_content @role.name
+    expect(page).to have_content 'super role'
+    search "super"
+    expect(page).to have_no_content @role.name
+    expect(page).to have_content 'super role'
+  end
+
   scenario "User creates a new role" do
     visit releaf_permissions_roles_path
     create_resource do